+++ src/main/java/com/progparcomposant/reseausocial/controllers/InvitationController.java
@@ -31,6 +31,17 @@ public class InvitationController {
         this.invitationConverter = invitationConverter;
     }
 
+    @GetMapping
+    public List<InvitationDTO> findAllInvitations() {
+        Iterable<Invitation> invitations = this.invitationRepository.findAll();
+
+        if (IterableUtils.size(invitations) > 0) {
+            return this.invitationConverter.entityToDto(IterableUtils.toList(invitations));
+        } else {
+            throw new NoSuchElementException(ErrorMessagesEnum.INVITATION_NO_INVITATIONS_IN_DATABASE.getErrorMessage());
+        }
+    }
+
     @PostMapping(path = "/new")
     public InvitationDTO newInvitation(@RequestBody InvitationDTO newInvitationDTO) {
         return invitationConverter.entityToDto(this.invitationRepository.save(invitationConverter.dtoToEntity(newInvitationDTO)));
@@ -85,11 +96,16 @@ public class InvitationController {
 
     @DeleteMapping(path = "/delete/{firstUserId}/{secondUserId}")
     public void deleteInvitationById(@PathVariable("firstUserId") Long firstUserId, @PathVariable("secondUserId") Long secondUserId) {
-        this.invitationRepository.deleteByFirstUserIdAndSecondUserId(firstUserId, secondUserId);
+        Optional<Invitation> invitation = this.invitationRepository.findByFirstUserIdAndSecondUserId(firstUserId, secondUserId);
+        if (invitation.isPresent()) {
+            this.invitationRepository.deleteByFirstUserIdAndSecondUserId(firstUserId, secondUserId);
+        } else {
+            throw new NoSuchElementException(ErrorMessagesEnum.INVITATION_NOT_FOUND.getErrorMessage());
+        }
     }
 
     @DeleteMapping(path = "/delete/all/{userId}")
     public void deleteAllInvitationsByUserId(@PathVariable("userId") Long userId) {
-        this.invitationRepository.deleteAllByFirstUserId(userId);
+        this.invitationRepository.deleteAllByFirstUserIdOrSecondUserId(userId, userId);
     }
 }
